requirements.txt
fastapi
uvicorn[standard]
redis
requests
python-dotenv
torch
pandas
scikit-learn
matplotlib
alpha_vantage
jinja2
python-multipart
ta
joblib
numpy
